server/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pathlib==1.0.1
pip==24.2
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==0.13.0
uuid==1.30
Werkzeug==3.0.4
groq==0.11.0